move ResourceUrlCreator and split out some code into UrlProber
