+++ core/src/main/java/net/adoptopenjdk/icedteaweb/resources/initializer/UnversionedResourceInitializer.java
@@ -3,7 +3,6 @@ package net.adoptopenjdk.icedteaweb.resources.initializer;
 import net.adoptopenjdk.icedteaweb.logging.Logger;
 import net.adoptopenjdk.icedteaweb.logging.LoggerFactory;
 import net.adoptopenjdk.icedteaweb.resources.Resource;
-import net.adoptopenjdk.icedteaweb.resources.ResourceUrlCreator;
 import net.adoptopenjdk.icedteaweb.resources.UrlRequestResult;
 import net.adoptopenjdk.icedteaweb.resources.cache.Cache;
 import net.adoptopenjdk.icedteaweb.resources.cache.ResourceInfo;
@@ -12,7 +11,7 @@ import java.net.URL;
 import java.util.ArrayList;
 import java.util.List;
 
-import static net.adoptopenjdk.icedteaweb.resources.ResourceUrlCreator.getUrl;
+import static net.adoptopenjdk.icedteaweb.resources.initializer.ResourceUrlCreator.getUrl;
 
 /**
  * Initializer for unversioned resources.
